apps/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-socketio==5.11.0
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.26.0
